Check data formats

Functions to check that election and prediction data are in the correct format and enforce assumptions about those data

Usage

check_elec_data(d, full = TRUE, strict = FALSE)

check_pred_data(d, strict = FALSE)

Arguments

d

A data frame to check.

full

If FALSE, omit some time-consuming checks of completeness.

strict

If TRUE, enforces that column names and order exactly match COLS_ELEC or COLS_PRED.

Value

TRUE, invisibly, if data are in correct format with assumptions met. Errors are informative where possible.

Functions

check_elec_data(): Check election data

check_pred_data(): Check prediction data

Schema

Election and prediction data schema, version 0.1

Tidy election data

For different analyses, election data often needs to be in a different format. The smallest unit of election data, which corresponds to a dataframe row in the "longest" format, consists of:

state, the uppercase USPS abbreviation of the state. The STATE_AP constant in the package can be used to map these abbreviations to Associated Press state abbreviations for display purposes.

election, a date for the particular election, stored in CSV as YYYY-MM-DD.

jurisdiction, a character identifier for the jurisdiction of the office. Use "US" for President, otherwise the state abbreviation, or a city or county name.

contest, a lowercase character identifier for the type of contest. Use "pres", "sen", "house", "gov", "stsen", "sthouse", etc (a full listing is in the CONTESTS variable). Suffix a contest with _dem and _rep for the Democratic and Republican primaries.

district, an positive integer identifier for the district number of the office. Use NA where not necessary, and 1 for at-large districts.

candidate, a character identifier for the candidate. Recommended format: "<last>, <first> [<middle>][, <suffix]" with proper capitalization. Can be NA where not necessary.

party, a lowercase character identifier for the party of the candidate. One of "dem" or "rep" or "ind" or "oth". The PARTIES constant in the package can be used to map these abbreviations to full party names.

time, a timestamp for the data record.

stage, the type of data. Recommended format is integer editorial stage: 0 for no vote, 2 for some vote, 4 for an end-of-night count, 6 for a final count pending provisional ballots, and 10 for a certified count. Alternatively, use "final", "partial", "not reported", etc.

method, the vote method reported. One of "eday", "early", "mail", "abs", "adv" (for combo of early, mail, or absentee), "prov", or "all". The VOTE_METHODS constant in the package can be used to map these abbreviations to full names.

geo, a unique geographic character identifier. Recommended: FIPS code for counties, GEOID or similar for precincts, "all" for race-wide data.

votes, a non-negative numeric value of the vote count.

Character vectors can also be factors. The value "all" is rserved to indicate aggregation over a column. Additional columns are permitted. Some recommended names for possible additional columns include:

past_*, for past vote data

flag, a character idnetifier for any quality control flags

Tidy prediction data

Model-based predictions at the geography or race level also have a standardized format. Columns below with the same name as election-data columns are formatted identically, i.e., a join should be possible on all shared columns.

state

election

jurisdiction

contest

district

candidate

party

pred_type, a character identifier for the type of predictions, e.g., "vote", "share", "margin_vote", or "margin_share". For margins, party and candidate should identify the specific subtraction, e.g. "rep - dem" for party. Republican minus Democratic should be the default.

time

model, a character identifier for the prediction model.

model_conf, a character identifier for the interval prediction method. The confidence or credible interval level should be clear from this field. Should be NA if no interval is provided.

pct_rep, a numeric value between 0 and 1 that proxies the percentage of reported units (e.g., precincts in).

rep, a numeric value of the reported vote count or share.

pred, a numeric value of the predicted vote count or share.

pred_min, a numeric value of the lower end of the predicted interval.

pred_max, a numeric value of the upper end of the predicted interval.

Assumptions

to be added as assumptions are checked.

Examples

check_elec_data(nj)