Bayesian linear regression

Efficient Bayesian linear regression with conjugate prior on the fitted values. Supports the usual S3 generics.

Usage

lm_bayes(
  formula,
  data,
  subset,
  weights,
  prior_yhat_loc = 0,
  prior_yhat_eff = 0.1,
  prior_sigma_scale = 0.05,
  prior_sigma_df = 4,
  ...
)

Arguments

formula

A formula specifying the model.

data

A data frame

subset

Optional subset of the data to use in fitting the model.

weights

Optional weights for each observation.

prior_yhat_loc

Prior mean vector for the predicted values.

prior_yhat_eff

Effective sample size for the prior on the predicted values.

prior_sigma_scale

Prior scale parameter for the error standard deviation. One can compute the implied median on the standard deviation with qinvchi().

prior_sigma_df

Prior degrees of freedom for the error standard deviation.

...

Additional arguments passed to model.frame and model.matrix.

Value

A fitted model object of class lm_bayes.