Legacy model configuration options

This is a helper function to specify configuration options for model_legacy. It does no work and exists only to aid in setup.

Usage

model_legacy_config(
  by_method = TRUE,
  strata = c("geostratum"),
  min_complete_per_stratum = 1L,
  min_unit_rep = 0.05,
  stage_complete = 6L,
  tr_pre = 1,
  tr_min = 0.25,
  tr_max = 4,
  check = TRUE
)

Arguments

by_method

If TRUE, estimate total vote and vote share separately by method.

strata

Character vector of column names that define strata for adjustment.

min_complete_per_stratum

Minimum number of complete observations (stage >= 6) required per stratum for adjustment to be applied.

min_unit_rep

Minimum proportion of total vote represented by a unit for it to be included in the vote share model.

stage_complete

Stage at which a county is considered complete for the total vote model.

tr_pre

Pre-election turnout ratio adjustment factor. The past vote will be multiplied by this factor.

tr_min, tr_max

Minimum and maximum bounds for the turnout ratio adjustment factor. Turnout ratios (ratio to past vote) outside these bounds will be truncated to the bounds.

check

If TRUE, check input data with check_elec_data().

Value

A list with configuration parameters.