Sample New York Mayoral Election Data

Sample New York Mayoral Election Data

Usage

nyc

Format

A data frame with 4,063 rows and 8 columns:

ad

Assembly District number

ed

Election District (precinct) identifier

obs

Time-of-night vector. Threshold to simulate whether a precinct has been observed.

mamdani

Zohran Mamdani vote share

prim25_cvr_top_m

Fraction of first-round ballots ranking Mamdani first in the primary election

votes_total_21

Total votes in the 2021 mayoral general election

p_white

Fraction of voters identifying as White alone

total

Total number of ballots cast in the current election

Examples

head(nyc)
#> # A tibble: 6 × 8
#>      ad ed      obs mamdani prim25_cvr_top_m votes_total_21 p_white total
#>   <dbl> <chr> <dbl>   <dbl>            <dbl>          <dbl>   <dbl> <dbl>
#> 1    77 001   0.851   0.662            0.354           63.1  0.114    414
#> 2    77 002   0.278   0.684            0.343          149.   0.113    513
#> 3    77 003   0.689   0.656            0.318          111.   0.0905   471
#> 4    77 004   0.467   0.734            0.632           16.4  0.232    109
#> 5    77 005   0.500   0.672            0.284          113.   0.0959   527
#> 6    77 006   0.370   0.678            0.232           84.2  0.185    510